requirements.txt
camelot-py[cv]
pdfplumber
ocrmypdf

pandas
openpyxl
xlsxwriter

requests
tesseract
ghostscript

certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
Deprecated==1.2.18
deprecation==2.1.0
et_xmlfile==2.0.0
idna==3.10
img2pdf==0.6.1
lxml==6.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pdfminer.six==20250506
pi_heif==1.1.0
pikepdf==9.10.2
pillow==11.3.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
pypdf==3.17.4
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2025.2
rich==14.1.0
setuptools==65.5.1
six==1.17.0
tabulate==0.9.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
wrapt==1.17.2